Adds the CDTDatastore+Conflicts -resolveConflictsForDocument code and tests.
The -resolveConflictsForDocument:resolver:error method is used by external developers
to facilitate conflict resolution. The caller of this method supplies a delegate object
to perform the resolution, while this method takes care to insert to the new revision
into the database and to ensure the document is no longer conflicted by adding a deleted
revision to the appropriate leaf nodes.

Adds DatastoreConflictResolvers.h and .m, which define example objects that
follow the CDTConflictResolver protocol and are used in these tests.

Adds the excludeDeleted: option to TD_Database -getAllRevisionsOfDocumentId. This
allows TD_Database to return all of the active current revisions for a document that
are not 'deleted'. This set of revisions are the confliction revisions of a document.