This is a temporary commit and will be squashed.
This adds a few tests, which currently fail, to the conflicts test suite.

This is the start of fixing the conflict resolution mechanism to support
attachments

In particular, I don't understand the sequence ordering of attachments.
testCreateConflict and testCreateConflictWithAttachment print out
a docs/revs JOINed table and the attachments table. I would expect the sequence
number for the attachment to be 2 and not 3 (in -testCreateConflictWithAttachment)
addenda: I discovered that compacting the database changes the sequence number
of the row in the attachments table. Is this a bug?

We should  discuss -testResolveConflictWithAttachmentForRev2b. I expect
that if I choose the 2-b revision to be the winner, which shouldn't have attachments,
that my new winning revision contains no attachment. However, printing the
tables to console shows that the attachments table gets a new row with sequence = 6,
the resolved doc.
addenda: This should be fixed with the proper handling of attachments