It seems that setting the log level in CloudantTests.m has no affect and that only LogWarns are being printed, which is the default log level set in CDTLogging.h
ChangeLogLevel must be called somewhere within an instance
method of a class in order to change the logging level for that
class.

So, in order to use this in a robust way, one would probably have
to wrap a ChangeLogLevel call inside a ifdef preprocessor macro,
or runtime arg/envioronment variable.