Adds potential fix to CDTLogging level issue.
Adds Log statements to testDictionaryForPushReplication test.

This doesn't work as expected for me. I initially expect the log
level "ALL" to be applied globally (via the +initialize class
method for CloudantTests.m). Instead, the log level seems to be
stuck at the default of 'Warn'.
Also, with the added Log statements in testDictionaryForPushReplication,
I see multiple statements printed to the console for each call.