Prototype replicator specific to pull replicating by docID
This replicator is mostly copy/paste from TDPuller, but it implements
a simpler pull-by-docId algorithm which just pulls all the documents
in a list, rather than trying to make use of _changes. For large
databases and small sets of document IDs this is potentially more
efficient.

The way this works shouldn't interfere with normal replications, as
document revisions are inserted as normal as they are found, just like
if there were two pull replications replicating documents from the
same database -- one might be surprised to find changes already in the
local database, but it shouldn't be a problem.

Right now it looks like this code only pulls the winning revision, but
I'd need to check that with a test.