Applies fix found in CBL.
Removes request to set remote checkpoint doc from list of
remote requests that are being tracked. The idea is that
when the replicator stops, this remote request isn't canceled.

However, this doesn't work and the tests fail.

Prevents setting the TDReplicator error when the error is for
a local database being deleted.