Removes the TDReplicator -clearDbRef method.
This method seemed to be appropriate when all replications
were executed on a main replication thread. Now, it could
interfere with proper synchronization of the last sequence
value betweent the local and remote checkpoint documents.

Additionally, the pointer to the TD_Database is now set to
nil after the replicator's thread has exited its runloop.
Setting the pointer to nil is still necessary since it's
the mechanism to make each TDReplicator object only
useable once.